moved boundary conditions, thermostats to respective files, prepared simulation from sheet 02d
