--- Exercise Sheet 02/code/solar_system.py
@@ -1,116 +1,0 @@
-import numpy as np
-
-solar_system_initial_configuration = np.array(
-    [
-        [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10],  # Particle ID
-        [
-            "Sun",
-            "Mercury",
-            "Venus",
-            "Earth",
-            "Mars",
-            "Jupiter",
-            "Saturn",
-            "Uranus",
-            "Neptun",
-            "Pluto",
-            "Ceres",
-        ],  # Type
-        [
-            0,
-            -8.305091745229271e-03,
-            -4.655560505865924e-01,
-            -8.594619144287726e-01,
-            -1.552939778499883e00,
-            2.105413944056840e-01,
-            9.515406434704190e00,
-            1.073262320911012e01,
-            2.987730734963971e01,
-            1.852952071857357e01,
-            2.724728113705883e00,
-        ],  # x position
-        [
-            0,
-            -4.676619114541750e-01,
-            -5.635653494750051e-01,
-            -5.344345587457319e-01,
-            6.082780616338184e-01,
-            5.110410644904475e00,
-            -1.158914726763215e00,
-            1.631315330407309e01,
-            -2.886321709780097e-01,
-            -2.990778423851285e01,
-            -1.086356480859120e00,
-        ],  # y position
-        [
-            0,
-            -3.731393296322649e-02,
-            1.907402983072339e-02,
-            2.040293677989342e-04,
-            5.098429859993224e-02,
-            -2.590503416060689e-02,
-            -3.587069951127249e-01,
-            -7.845649796549910e-02,
-            -6.826089382944749e-01,
-            -2.159516183836656e00,
-            -5.369223277851006e-01,
-        ],  # z position
-        [
-            0,
-            2.249752881715944e-02,
-            1.546706410013126e-02,
-            8.781678971499092e-03,
-            -4.624728987586821e-03,
-            -7.625989566072758e-03,
-            3.651726124137276e-04,
-            -3.314709667922642e-03,
-            1.001321809415393e-05,
-            2.761591067905998e-03,
-            3.396432983254987e-03,
-        ],  # vx velocity
-        [
-            0,
-            1.233724858277215e-03,
-            -1.296876647045673e-02,
-            -1.469852019614368e-02,
-            -1.181591874295253e-02,
-            6.704904690423375e-04,
-            5.525515091947649e-03,
-            1.978428597105292e-03,
-            3.157565709409927e-03,
-            9.598823007529163e-04,
-            8.925383520139937e-03,
-        ],  # vy velocity
-        [
-            0,
-            -1.961938017056361e-03,
-            -1.070291132368980e-03,
-            1.945103428955565e-07,
-            -1.341074095976234e-04,
-            1.679103446738440e-04,
-            -1.102899088441146e-04,
-            5.041122823830823e-05,
-            -6.560401788042341e-05,
-            -9.104358899121763e-04,
-            -3.419560192091928e-04,
-        ],  # vz velocity
-        [0] * 11,  # fx force
-        [0] * 11,  # fy force
-        [0] * 11,  # fz force
-        [
-            19.861,
-            0.3302e-5,
-            4.8685e-5,
-            5.9736e-5,
-            6.4185e-6,
-            1.8986e-2,
-            5.68429e-3,
-            8.6832e-4,
-            1.0243e-3,
-            0.0127e-5,
-            9.3930e-9,
-        ],  # mass
-        [0.2] * 11,  # radius
-    ],
-    dtype=object,
-)
